requirements.txt
torch>=1.7.0
torchvision>=0.8.1
streamlit>=1.10.0
numpy>=1.19.2
pandas>=1.1.3
matplotlib>=3.3.2
seaborn>=0.11.0
plotly>=5.3.1
scikit-learn>=0.23.2
onnx>=1.10.0
onnxruntime>=1.8.0
Pillow>=8.0.1
tqdm>=4.50.2
opencv-python>=4.4.0.44
